*hannes: fixed syntax errors
git-svn-id: file:///usr/local/Singular/svn/trunk@2670 2c84dea3-7e68-4137-9b89-c4e89433aadc

+++ MP/MPT/GP_IsOk.cc
@@ -359,7 +359,7 @@ bool GP_DistMvPoly_t::IsDistMvPolyDataOk(const void* data)
     }
     else
     {
-      int* evector = NULL;
+      long int* evector = NULL;
       ExpVector(monom, evector);
       
       if (evector == NULL) return NULL;
